Move enumerable to collections.
