+++ tests/xharness/Jenkins/TestTasks/RunSimulatorTask.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using xharness.Collections;
 using xharness.Hardware;
 
 namespace Xharness.Jenkins.TestTasks
