+++ tests/xharness/Jenkins/TestTasks/RunXITask.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using xharness.Collections;
 using xharness.Hardware;
 using Xharness.Logging;
